Add files via upload
The Smart Trash Bin is an innovative IoT-based solution designed to enhance waste management efficiency. It uses sensors to detect the fill level of the bin and automatically notify municipal authorities when the bin is full. The system integrates a microcontroller, ultrasonic sensors, and a Wi-Fi module to enable real-time monitoring and communication. This project aims to reduce overflow issues, improve cleanliness, and promote timely waste collection, thereby contributing to a smarter and cleaner environment.